Week 1 Notebook_full
Full version of of data cleaning and aggregation where there are tools for generating url for data (both on local machine and in the Internet), automated cleaning, filtering and aggregating data. Pivot tables are collected in a child folder and stored in separate files for each processed month. Although I managed to process 24 files in about 50 minutes (each ~1,7 GiB, so all data is about 40 GiB), it is still more efficient to save aggregated data for further analysis.
Still only one zone and one month is visualised